Routex.Attrs (Phoenix Routes Extension Framework v0.2.0-alpha.7)

Provides an interface to access and update Routex attributes.

Extensions can make use of Routex.Attrs values provided by Routex itself, Routex backends and other extensions. As those values are attributes to a route extension B can use values attributed to to a route by (pre)processing extension A.

To make the availability of the attributes as predictable as possible, Routex uses a flat structure.

Extension developers are encouraged to put as much information into the attributes as possible.

Extensions should add any fallback/default they might use themselves to the attributes.

Functions

cleanup(route)

Removes non private fields from attrs. Input is either an opts map or a Route.

get(route, key \\ nil, default \\ nil)

Get the value assigned to key from route's options or return default. When no key is provided, the function returns all options.

get!(route, key, error_msg \\ nil)

Get the value assigned to key from route's options or raise if key not found. Accepts an optional custom error_msg as third argument

is_private(key)

@spec is_private({atom(), any()} | atom()) :: boolean()

Returns true when the provided key or attribute is private.

merge(route, value)

Merge value into route's options.

put(route, value)

Assigns value to route's options.

put(route, key, value)

Assigns value to key in route's options.

update(route, fun)

Updates the options of a route by applying given function.

update(route, key, fun)

Updates the value assigned to key of a route's options by applying given function.