Routex.Route (Phoenix Routes Extension Framework v0.2.0-alpha.7)

Function for working with Phoenix Routes

Functions

exprs(route, env)

@spec exprs(Phoenix.Router.Route.t(), Macro.Env.t()) :: map()

Compatibility wrapper around Phoenix.Router.Route.exprs

get_nesting(route, offset \\ 0)

Returns the nesting level of an (ancestor) route. By default the parent. This can be adjusted by providing an negative depth offset.

group_by_method_and_path(routes, offset \\ 0)

Returns routes grouped by the combination of method and path of an (ancestor) route. By default groups by parent. This can be adjusted by providing a negative depth offset.

group_by_nesting(routes, offset \\ 0)

Returns routes grouped by nesting level of an (ancestor) route. By default groups by parent. This can be adjusted by providing an negative depth offset